Moving helper functions into the files they're only used in.
